requirements.txt
crewai>=0.28.0
ollama>=0.1.7
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
